Rename types.rs to path_update.rs
