+++ src/cli/self_update/path_update.rs
